cosmetic: move findMethodOnAppFromData to utils/apps
