+++ packages/aragon-wrapper/src/utils/index.js
@@ -1,4 +1,3 @@
-import { keccak256 } from 'js-sha3'
 import { isAddress } from 'web3-utils'
 import ContractProxy from '../core/proxy'
 import { getAbi } from '../interfaces'
@@ -47,24 +46,6 @@ export function makeProxy (address, interfaceName, web3, initializationBlock) {
 
 export function makeProxyFromABI (address, abi, web3, initializationBlock) {
   return new ContractProxy(address, abi, web3, initializationBlock)
-}
-
-/**
- * Find the method descriptor corresponding to the data component of a
- * transaction sent to `app`.
- *
- * @param  {Object} data Data component of a transaction to app
- * @param  {Object} app App artifact
- * @return {Object} Method with radspec notice and function signature
- */
-export function findMethodOnAppFromData (data, app) {
-  if (app && app.functions) {
-    // Find the method
-    const methodId = data.substring(2, 10)
-    return app.functions.find(
-      (method) => keccak256(method.sig).substring(0, 8) === methodId
-    )
-  }
 }
 
 export { default as AsyncRequestCache } from './AsyncRequestCache'
